backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
pymysql==1.1.1
python-dotenv==1.0.1
pydantic==2.9.0
pydantic-settings==2.5.0
twilio==9.3.2
cryptography==43.0.1
